requirements.txt
fastapi
uvicorn
pydantic
pydantic_ai
PyYAML
Pillow
opencv-python
ffmpeg-python
numpy
